Update README.md
This is a modification of the AutoHotkey Script seen at  https://www.computerhope.com/tips/tip224.htm which uses keyboard hotkeys to add, remove and switch between desktops. This version strives to do the same with a clickable GUI positioned at the top or bottom of the screen (which is especially useful for tablets).